create FilterPopover
